requirements.txt
# FastAPI Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# AI / RAG
langchain==0.3.0
langchain-community==0.3.0
sentence-transformers==3.0.0  # chromadb removed (we use FAISS-based store instead)

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Web Crawling
beautifulsoup4==4.12.3
requests==2.32.0
lxml==5.3.0
fake-useragent==1.5.1

# Knowledge Graph
networkx==3.3

# Utilities
httpx==0.27.0
python-dotenv==1.0.1
APScheduler==3.10.4
